CI/Release: Allow overriding mvn goal and profile
This patch allows the `mvn` goal and profile of the mavenCI and
mavenCanaryRelease steps to be overridden. User can override the
entire command by setting the `cmd` variable or the goal and the
profile by setting `goal` and `profile` respectively.

See ReadMe.md for details

Related to: https://github.com/openshiftio/openshift.io/issues/4151