Add new config api
The config api stores global settings of a pipeline. The initial
version adds support for accessing `runtime` parameter which is now
used in `build` to spawn the container that has the build tools.

e.g setting `config runtime: 'node'`, build and deploy will use
the `node` container instead of the default `oc`